Added GUI and Event changes
Added play and stop buttons. The play button has basic functionality.
The stop button is hooked up for adding functional behavior. The play
functionality was initially triggered when the "GO!" button was clicked,
but the "GO!" button was actually for creating a new .wav file with the
secret encrypt message. So I moved it to the play button's behavior.