Added Comments and Split Function into 2.
Added comments throughout the file. Added the function
"setWavFileToNotValid" to indicate when a WAV file has failed to
initialize and is not a valid WAV file. The getter function
"hasValidWavFile" is completed and so I deleted the "TO-DO" comments for
it. Split the "createWavFileAndInjectMessage" into two different
operations: 1) create a copy of the WAV file into a separate JWav
object, and 2) have the copied JWav object inject the encrypted message
by calling the newly created function "injectEncryptedMessage".