Added Comments, Variables, Functions
Added comments throughout the file. Added a copy constructor. Added
variables and functions that will check if the WAV file is valid, and if
the WAV file contains an injected message. Added "displayWavFileInfo"
function to display useful information about the WAV file for debug
purposes. Split the job of "createWavFileAndInjectMessage" function in 2
where the copy constructor and the new function "injectEncryptedMessage"
will replace "createWavFileAndInjectMessage". Added validation code to
the "play" and "stop" functions. Added code to "play" function where the
audio file will be able to replay from the beginning of the file when
the file ends and the user clicks the play button again.