Fixed Bug, Added More Functionality
Fixed a resource leak where audio streams weren't being closed when the
user was finished with it. Added a test button for testing purposes.
Rearranged the order of how the Encryption process will run. Now the
copied WAV file has no communication with the original WAV file after
the initial copy. Before the original WAV file was going to perform the
encrypting and then pass that data to the copied WAV file.