Pulled Code from Master Branch
Pulled code from the Master Branch to this branch. This branch will be
to create a new GUI setup where the user will only have two modes to
choose from: Encrypt, Decrypt. In effect, the user will only see options
for one of the modes at a time. This should make for an easier UX, and a
cleaner GUI.