Refactored GUI and Added Validation Code
The GUI is now more user friendly and pleasing to the eye. Added
validation code for "goButton", "playButton", and "stopButton" so that
they are only enabled when the necessary input is given. Added a reset
button as well that resets the whole form. Added a new class
"MyDocumentListener" that helps in validating the password and message
fields.