Refactored and Added Validation/Pause Functionality
Refactored some of the code. Made the play button into a play/pause
toggle button. The goButton now stops and resets the WAV file's audio
(before it acted more like a pause button). Added validation code for
enabling/disabling parts of the GUI (goButton, playPauseButton,
stopButton). Added code to display a dialog window whenever there is an
error or the encryption process is a success.