Fixed Bug, and Added Extra Security Measure
Fixed a bug where whenever the user played a wav file in decrypt mode,
then changed the mode from decrypt to encrypt, and tried to inject a
message into the wav file previously played in decrypt mode, an error
would say that the wav file was being used in another process. I also
made it so that 300 characters are always displayed when decrypting
(padded the real message to display with spaces when the correct
password is entered).